refactor: move left-and-right structures to service
Currently, LeftAndRightTimeStructure and LeftAndRightLockStructure
are defined in trainrun-section-tab.component.ts. The component is
imported in various places just to grab these two types.

Define the types in the service to avoid importing components from
services.
