+++ src/app/services/data/trainrun-section-times.service.ts
@@ -1,16 +1,26 @@
 import {Injectable} from "@angular/core";
 import {MathUtils} from "../../utils/math";
 import {LeftAndRightElement, TrainrunsectionHelper} from "../util/trainrunsection.helper";
-import {
-  LeftAndRightLockStructure,
-  LeftAndRightTimeStructure,
-} from "../../view/dialogs/trainrun-and-section-dialog/trainrunsection-tab/trainrun-section-tab.component";
 import {TrainrunService} from "./trainrun.service";
 import {TrainrunSectionService} from "./trainrunsection.service";
 import {FilterService} from "../ui/filter.service";
 import {TrainrunSection} from "../../models/trainrunsection.model";
 import {Node} from "../../models/node.model";
 import {LoadPerlenketteService} from "../../perlenkette/service/load-perlenkette.service";
+
+export interface LeftAndRightTimeStructure {
+  leftDepartureTime: number;
+  leftArrivalTime: number;
+  rightDepartureTime: number;
+  rightArrivalTime: number;
+  travelTime: number;
+}
+
+export interface LeftAndRightLockStructure {
+  leftLock: boolean;
+  rightLock: boolean;
+  travelTimeLock: boolean;
+}
 
 @Injectable({
   providedIn: "root",
